feat: Implement InternalError helper
