+++ wrap_details_test.go
